[vpc] Fix VPC sources build on Windows (at least)

+++ external/vpc/public/mathlib/mathlib.h
@@ -326,7 +326,7 @@ qboolean VectorsEqual( const float *v1, const float *v2 );
 
 inline vec_t RoundInt (vec_t in)
 {
-	return floor(in + 0.5f);
+	return floorf(in + 0.5f);
 }
 
 size_t Q_log2( unsigned int val );
@@ -1356,7 +1356,7 @@ FORCEINLINE unsigned long RoundFloatToUnsignedLong(float f)
 
 FORCEINLINE bool IsIntegralValue( float flValue, float flTolerance = 0.001f )
 {
-	return fabs( RoundFloatToInt( flValue ) - flValue ) < flTolerance;
+	return fabsf( RoundFloatToInt( flValue ) - flValue ) < flTolerance;
 }
 
 // Fast, accurate ftol:
@@ -2185,7 +2185,7 @@ FORCEINLINE float * UnpackNormal_UBYTE4( const unsigned int *pPackedNormal, floa
 	y = ( y*ySign - ySignBit ) / 63.0f;
 	z = 1.0f - x - y;
 
-	float oolen	 = 1.0f / sqrt( x*x + y*y + z*z );	// Normalize and
+	float oolen	 = 1.0f / sqrtf( x*x + y*y + z*z );	// Normalize and
 	x			*= oolen * xSign;					// Recover signs
 	y			*= oolen * ySign;
 	z			*= oolen * zSign;
@@ -2346,14 +2346,14 @@ inline float FastPow10( float i ) { return exp2f( i * LOGBASE2OF10 ); }			// 10^
 
 inline bool CloseEnough( float a, float b, float epsilon = EQUAL_EPSILON )
 {
-	return fabs( a - b ) <= epsilon;
+	return fabsf( a - b ) <= epsilon;
 }
 
 inline bool CloseEnough( const Vector &a, const Vector &b, float epsilon = EQUAL_EPSILON )
 {
-	return fabs( a.x - b.x ) <= epsilon &&
-		fabs( a.y - b.y ) <= epsilon &&
-		fabs( a.z - b.z ) <= epsilon;
+	return fabsf( a.x - b.x ) <= epsilon &&
+		fabsf( a.y - b.y ) <= epsilon &&
+		fabsf( a.z - b.z ) <= epsilon;
 }
 
 // Fast compare
